+++ src/services/file/index.js
@@ -1,7 +1,5 @@
 const { FileService } = require('./file.service')
-const { TargetFileAlreadyExist } = require('./target-file-already-exist.error')
 
 module.exports = exports = {
-  FileService,
-  TargetFileAlreadyExist
+  FileService
 }
